requirements.txt
fastapi[standard]>=0.115,<1.0
pydantic>=2.0,<3.0
google-cloud-secret-manager>=2.20,<3.0
google-cloud-storage>=2.18,<4.0
google-genai>=0.8,<2.0
python-dotenv>=1.0,<2.0
motor>=3.5,<4.0
pymongo>=4.9,<5.0
bcrypt>=4.2,<6.0
redis>=5.0,<6.0
httpx>=0.27,<1.0
